refactor: set aside objects
